pyproject.toml
[project]
name = "SmudgeLord"
version = "1.0.0"
description = "SmudgeLord is a bot developed in Python with some fun and useful functions."
authors = [{ name = "Luiz Renato", email = "ruizlenato@proton.me" }]
readme = "README.rst"
license = { file = "LICENSE" }
requires-python = ">=3.12"
dependencies = [
    "hydrogram @ git+https://github.com/hydrogram/hydrogram@dev",
    "tgcrypto>=1.2.5",
    "uvloop>=0.19.0",
    "aiosqlite>=0.19.0",
    "yt-dlp>=2023.10.13",
    "httpx[http2, socks]>=0.19.0",
    "babel>=2.13.1",
    "filetype>=1.2.0",
    "humanize>=4.8.0",
    "beautifulsoup4>=4.12.2",
    "esprima>=4.0.1",
    "gpytranslate>=1.5.1",
    "pyaml>=23.9.7",
    "sentry-sdk>=1.34.0",
    "pillow>=10.1.0",
]


[project.urls]
Repository = "https://github.com/ruizlenato/SmudgeLord/"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = ["pre-commit>=3.6.0", "ruff>=0.1.11"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.ruff]
line-length = 99
show-fixes = true
target-version = "py312"
select = [
  "I", # isort
  "E", # pycodestyle
  "F", # pyflakes
  "UP",  # pyupgrade
  "PTH",  # flake8-use-pathlib
  "C4",  # flake8-comprehensions
  "SIM", # flake8-simplify
  "RET", # flake8-return
  "RUF",  # ruff
]

[tool.ruff.isort]
known-first-party = ["smudge"]

[tool.hatch.build.targets.wheel]
packages = ["smudge"]